refactor(search): Move order-by.ts into SearchTracePage
model/order-by.ts had exactly one consumer: the SearchTracePage
feature (its sort dropdown, table sort, and search summary sorting).
Nothing else in the app imports it. Keeping it in the shared model/
directory implied broader applicability that doesn't exist, and
didn't match the codebase's own convention of colocating
feature-scoped logic inside its feature directory (e.g.
TracePage/types.ts, SearchTracePage/search-panel-store.ts).

Moved order-by.ts and its test to components/SearchTracePage/ and
updated the 8 files that import from it. model/search.ts now reaches
into components/SearchTracePage/order-by since it has the same
single-consumer scope but wasn't asked to move.
