+++ packages/jaeger-ui/src/components/SearchTracePage/SearchResults/TraceTable.test.tsx
@@ -6,8 +6,8 @@ import { render, screen, fireEvent, within } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import { vi, describe, it, expect, beforeEach } from 'vitest';
 import TraceTable from './TraceTable';
-import * as orderBy from '../../../model/order-by';
-import type { OrderBy } from '../../../model/order-by';
+import * as orderBy from '../order-by';
+import type { OrderBy } from '../order-by';
 import { toOrderBy, fromOrderBy } from '../../../model/search';
 import type { Microseconds } from '../../../types/units';
 import { useSearchResultsStore } from '../store.search-results';
